回転の計算式を別ファイルに分けた｡1ファイルのコード数を減らすため
