+++ src/Rotate.js
@@ -1,3 +1,5 @@
+import RotateCalculations from "./RotateCalculations";
+
 export default class Rotate {
 
     constructor({
@@ -9,6 +11,7 @@ export default class Rotate {
         this.fieldWidth   = fieldWidth - 1
         this.fieldHeight  = fieldHeight - 1;
         this.checkCanMove = checkCanMove
+        this.rotateCalculations = new RotateCalculations()
     }
 
     
@@ -34,7 +37,7 @@ export default class Rotate {
         if (direction == "clockwise" ) {
             // 回したのを代入
             tentativeCoordinate = tetrimino.Coordinate.map(block => {
-                return this.clockwise({
+                return this.rotateCalculations.clockwise({
                     rotationPoint:tetrimino.Coordinate[tetrimino.clockwiseAxis],
                     beforeRotation:block
                 })
@@ -57,7 +60,7 @@ export default class Rotate {
 
         } else {
             tentativeCoordinate = tetrimino.Coordinate.map(block => {
-                return this.counterClockwise({
+                return this.rotateCalculations.counterClockwise({
                     rotationPoint:tetrimino.Coordinate[tetrimino.counterClockwiseAxis],
                     beforeRotation:block
                 })
@@ -83,72 +86,6 @@ export default class Rotate {
         tetrimino.Coordinate = tentativeCoordinate
 
         return tetrimino
-    }
-
-    /** 回転前点{a,b} 回転点 {c,d} */
-    /** X = (a - c)cos𝜃 - (b - d)sin𝜃 + c */
-    /** Y = (a - c)sin𝜃 + (b - d)cos𝜃 + d */
-
-    // Math.cosとかはなんかややこしいし､今回は90度しか使わないので使わない
-    /** cos90 = 0 
-     *  SIN90 = 1
-     */
-
-    /**
-     * 時計回り
-     * @param rotationPoint 回転軸
-     * @param beforeRotation 回転前の状態
-     * @returns 
-     */
-    clockwise({
-        rotationPoint,
-        beforeRotation
-    }){
-        
-        let afterRotation = {x:0,y:0}
-        afterRotation.x = 
-            (beforeRotation.x - rotationPoint.x) * 0
-            -
-            (beforeRotation.y - rotationPoint.y) * 1
-            + 
-            rotationPoint.x
-
-        afterRotation.y =
-            (beforeRotation.x - rotationPoint.x) * 1
-            +
-            (beforeRotation.y - rotationPoint.y) * 0
-            +
-            rotationPoint.y
-
-        return afterRotation
-    }
-
-    /**
-     * 反時計回り
-     * @param rotationPoint 回転軸
-     * @param beforeRotation 回転前の状態
-     * @returns 
-     */
-    counterClockwise({
-        rotationPoint,
-        beforeRotation
-    }){
-        let afterRotation = {x:0,y:0}
-        afterRotation.x = 
-            (beforeRotation.x - rotationPoint.x) * 0
-            -
-            (beforeRotation.y - rotationPoint.y) * -1
-            + 
-            rotationPoint.x
-
-        afterRotation.y =
-            (beforeRotation.x - rotationPoint.x) * -1
-            +
-            (beforeRotation.y - rotationPoint.y) * 0
-            +
-            rotationPoint.y
-
-        return afterRotation
     }
 
     /** すでにおいているぶつかるようだったらちょっと移動する
